pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tsad-orchestra"
version = "0.1.0"
description = "Agent for time series anomaly detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LLM provider
    "openai>=1.30.0",
    # LangChain / MCP agent stack
    "langchain>=1.2.16",
    "langchain-openai>=1.2.1",
    "langchain-mcp-adapters>=0.1.7",
    "fastmcp>=3.2.4",
    "mcp>=1.27.0",
    # Numerical & data
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    # Utilities
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "psycopg2-binary>=2.9.12",
    "pyod>=3.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
    "mypy>=1.10.0",
    "interrogate>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
strict = true
ignore_missing_imports = true


